pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.22.0",
    "python-telegram-bot>=22.1",
    "sift-stack-py>=0.7.0",
    "telegram>=0.0.1",
]
